index.html

<!doctype html>
<meta charset=utf-8>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
 <!-- Script for Loading various other resources -->
<script src="./resources//sample-Script.js"></script>
</head>

<body>
  <h1> 
    The style to this text is applioed from stylesheet inserted by indirectly loaded script. 
  </h1>
  <script>
        test(() => {
      const resources = performance.getEntriesByType("resource");
      let fetchID_is_implemented =false;
      if(resource[0].fetchID != undefined )
      fetchID_is_implemented = true;      
      assert_true(fetchID_is_implemented);
    }, "Ensure fetchID attribute is defined for entry of Resource TIming Object");

    test(() => {
      const resources = performance.getEntriesByType("resource");
      let initiator_is_implemented =false;
      if(resource[0].initiator != undefined )
      initiator_is_implemented = true;      
      assert_true(initiator_is_implemented);
    }, "Ensure initiator attribute is defined for entry of Resource TIming Object");

    test(() => {
      const resources = performance.getEntriesByType("resource");
      //The first Two entries are for the Testharnessjs files .
      const fetchID_of_parent_script = resource[2].fetchID;
      const initiator_fetchID_of_indirecty_loaded_script = resource[3].initiator;
      assert_equals( fetchID_of_parent_script, initiator_fetchID_of_indirecty_loaded_script);
  }, "Ensure initiator Attribute matches with the fetch ID of parent script");

  test(() => {
      const resources = performance.getEntriesByType("resource");
      //The first Two entries are for the Testharnessjs files .
      const fetchID_of_parent_script_for_stylesheet = resource[3].fetchID;
      const initiator_fetchID_of_indirecty_loaded_stylesheet = resource[4].initiator;
      assert_equals( fetchID_of_parent_script, initiator_fetchID_of_indirecty_loaded_script);
  }, "Ensure initiator Attribute of stylesheet matches with the fetch ID of parent        script");

  </script>
</body>
